+++ app/src/main/java/dev/arkbuilders/navigator/analytics/folders/impl/FoldersAnalyticsImpl.kt
@@ -1,7 +1,8 @@
-package dev.arkbuilders.navigator.analytics.folders
+package dev.arkbuilders.navigator.analytics.folders.impl
 
 import android.content.Context
 import dev.arkbuilders.navigator.R
+import dev.arkbuilders.navigator.analytics.folders.FoldersAnalytics
 import org.matomo.sdk.Tracker
 import org.matomo.sdk.extra.TrackHelper
 import javax.inject.Inject
